Move openSiblingPlotPanel function to separate file
Calling panels don't need to import all of the plot code to get this one function.
