+++ packages/studio-base/src/panels/RawMessages/Value.tsx
@@ -15,7 +15,8 @@ import { withStyles, makeStyles } from "tss-react/mui";
 
 import HoverableIconButton from "@foxglove/studio-base/components/HoverableIconButton";
 import Stack from "@foxglove/studio-base/components/Stack";
-import { openSiblingPlotPanel, plotableRosTypes } from "@foxglove/studio-base/panels/Plot";
+import { openSiblingPlotPanel } from "@foxglove/studio-base/panels/Plot/openSiblingPlotPanel";
+import { plotableRosTypes } from "@foxglove/studio-base/panels/Plot/types";
 import {
   openSiblingStateTransitionsPanel,
   transitionableRosTypes,
